+++ src/test/main.cpp
@@ -629,5 +629,5 @@ int main(int argc, char* argv[])
   auto numPassed = count_if(rs.begin(), rs.end(),
                             [] (const testinator::Result& r) { return r.m_success; });
   auto total = static_cast<decltype(numPassed)>(rs.size());
-  return total - numPassed;
+  return static_cast<int>(total - numPassed);
 }
